mods

+++ app/Http/Controllers/Admin/CoursesController.php
@@ -209,6 +209,10 @@ class CoursesController extends Controller
 
         $courses = Courses::find($id);
 
+        if ($courses->users()->count() > 0 ||  $courses->groups()->count() > 0) {
+            return redirect()->back()->with('error', __('pages.delete-failed-has-users'));
+        }
+
         $check = $courses->delete();
 
         if ($check) {
